+++ developer-extension/src/panel/components/tabs/eventsTab/index.ts
@@ -1,0 +1,1 @@
+export { EventsTab } from './eventsTab'
